refactor : recoil/theme, recoil/state 삭제 recoil/theme -> types/theme으로 변경 recoil/state state들의 호출이 잦지 않아 각각의 파일에서 import하도록 변경
