+++ src/component/Common/HeaderLayout/Drawer.tsx
@@ -14,8 +14,8 @@ import Inquiry from '../Icon/drawer/Inquiry';
 import DrawerItem from './DrawerItem';
 import DropList from './DropList';
 import { useLogin } from '@/hooks/useLogin';
+import { drawerState } from '@/recoil/drawer';
 import { titleFontState } from '@/recoil/font/title';
-import { drawerState } from '@/recoil/state';
 import { drawerBoard } from '@/types/drawerBoard';
 import { axiosGetUserBoard } from '@/utils/apiInterface';
 import { handleBackground } from '@/utils/handleBackground';
