+++ src/component/Common/HeaderLayout/TitleHeader.tsx
@@ -8,7 +8,8 @@ import Close from '../Icon/Close';
 import { HamburgerMenu } from '../Icon/Header';
 import useModal from '@/hooks/useModal';
 import { useScroll } from '@/hooks/useScroll';
-import { drawerState, titleState } from '@/recoil/state';
+import { drawerState } from '@/recoil/drawer';
+import { titleState } from '@/recoil/title';
 import { confirm } from '@/utils/confirm';
 import { drawerOpen } from '@/utils/drawerOpen';
 import { defaultState } from '@/utils/theme/default';
