+++ src/component/Common/StrcatBoard.tsx
@@ -4,7 +4,7 @@ import { useRecoilState } from 'recoil';
 
 import ObserveContent from './ObserveContent';
 import useModal from '@/hooks/useModal';
-import { addContentState } from '@/recoil/content';
+import { addContentState } from '@/recoil/addContent';
 import { board } from '@/types/boards';
 import { content } from '@/types/content';
 import { observeContent } from '@/types/observe';
