+++ src/recoil/addContent.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const addContentState = atom({
+  key: 'addContentState',
+  default: 0,
+});
