--- src/recoil/content.ts
@@ -1,16 +1,0 @@
-import { atom } from 'recoil';
-
-export const observeState = atom({
-  key: 'observeState',
-  default: {
-    boardId: '',
-    contentId: 0,
-    photoUrl: '',
-    writer: '',
-  },
-});
-
-export const addContentState = atom({
-  key: 'addContentState',
-  default: 0,
-});
